+++ lib/td_cache/templates/field_formatter.ex
@@ -19,6 +19,11 @@ defmodule TdCache.Templates.FieldFormatter do
     apply_role_meta(field, claims, role_name, user_roles)
   end
 
+  def format(%{"type" => "group", "values" => %{"role_groups" => role_name}} = field, ctx) do
+    user_group_roles = Map.get(ctx, :user_group_roles, %{})
+    apply_user_group_meta(field, role_name, user_group_roles)
+  end
+
   def format(%{"type" => "user_group", "values" => %{"role_groups" => role_name}} = field, ctx) do
     claims = Map.get(ctx, :claims, nil)
     user_roles = Map.get(ctx, :user_roles, %{})
@@ -77,10 +82,15 @@ defmodule TdCache.Templates.FieldFormatter do
        )
        when not is_nil(role_name) do
     groups = Map.get(user_group_roles, role_name, [])
-    names = Enum.map(groups, & &1.name)
+    names = Enum.map(groups, &group_name_or_alias/1)
     values = Map.put(values, "processed_groups", names)
     Map.put(field, "values", values)
   end
 
   defp apply_user_group_meta(field, _role, _user_roles), do: field
+
+  defp group_name_or_alias(%{alias: group_alias, name: name}) when group_alias in [nil, ""],
+    do: name
+
+  defp group_name_or_alias(%{alias: group_alias}), do: group_alias
 end