+++ test/td_cache/templates/preprocessor_test.exs
@@ -52,7 +52,7 @@ defmodule TdCache.Templates.PreprocessorTest do
     test "preprocess_template/2 enriches user_group role fields" do
       %{id: domain_id} = CacheHelpers.insert_domain()
       %{id: user_id, full_name: full_name} = CacheHelpers.insert_user()
-      %{id: group_id, name: group_name} = CacheHelpers.insert_group()
+      %{id: group_id, alias: group_alias} = CacheHelpers.insert_group()
 
       AclCache.set_acl_roles("domain", domain_id, [@role_name])
       AclCache.set_acl_group_roles("domain", domain_id, [@role_name])
@@ -83,7 +83,7 @@ defmodule TdCache.Templates.PreprocessorTest do
                "values" => %{
                  "role_groups" => @role_name,
                  "processed_users" => [full_name],
-                 "processed_groups" => [group_name]
+                 "processed_groups" => [group_alias]
                }
              }
 
@@ -131,7 +131,7 @@ defmodule TdCache.Templates.PreprocessorTest do
     test "preprocess_template/2 process dynamic table type fields" do
       %{id: domain_id} = CacheHelpers.insert_domain()
       %{id: user_id, full_name: full_name} = CacheHelpers.insert_user()
-      %{id: group_id, name: group_name} = CacheHelpers.insert_group()
+      %{id: group_id, alias: group_alias} = CacheHelpers.insert_group()
 
       AclCache.set_acl_roles("domain", domain_id, [@role_name])
       AclCache.set_acl_group_roles("domain", domain_id, [@role_name])
@@ -206,9 +206,44 @@ defmodule TdCache.Templates.PreprocessorTest do
                "name" => "user_group_field_col",
                "type" => "user_group",
                "values" => %{
-                 "processed_groups" => [group_name],
+                 "processed_groups" => [group_alias],
                  "processed_users" => [full_name],
                  "role_groups" => "foo_role"
+               }
+             }
+    end
+
+    test "preprocess_template/2 enriches group role fields with groups only" do
+      %{id: domain_id} = CacheHelpers.insert_domain()
+      %{id: user_id} = CacheHelpers.insert_user()
+      %{id: group_id, alias: group_alias} = CacheHelpers.insert_group()
+
+      AclCache.set_acl_roles("domain", domain_id, [@role_name])
+      AclCache.set_acl_group_roles("domain", domain_id, [@role_name])
+      AclCache.set_acl_role_users("domain", domain_id, @role_name, [user_id])
+      AclCache.set_acl_role_groups("domain", domain_id, @role_name, [group_id])
+
+      ctx = %{domain_ids: [domain_id], claims: %{user_id: user_id}}
+
+      fields = [
+        %{
+          "name" => "group_field",
+          "type" => "group",
+          "values" => %{"role_groups" => @role_name}
+        }
+      ]
+
+      template = %{content: [%{"name" => "group1", "fields" => fields}]}
+
+      assert %{content: [%{"fields" => [group_field]}]} =
+               Preprocessor.preprocess_template(template, ctx)
+
+      assert group_field == %{
+               "name" => "group_field",
+               "type" => "group",
+               "values" => %{
+                 "role_groups" => @role_name,
+                 "processed_groups" => [group_alias]
                }
              }
     end